Revert "layer-tap Lower/Tab, Lower/BS, Adjust/Del"
This reverts commit 0c678862b8b0c3cb3d0f0b9205b7b6df0b1f20bd.

It did not work very well. It turns out the outer thumb keys are not comfortable
to tap frequently as backspace key; and it is harder to repeat.
Prefer faster layer switch than having more editing keys on the thumb row.
Removed the TAPPING_FORCE_HOLD option; it's still desirable to repeat the rest
dual-role keys.